HCAP-1189 | Things are communicating properly now, next up refactor prettifyStatus -> components/ParticipantStatus
